requirements.txt
streamlit
langchain
langchain-community
langchain-openai
pypdf
faiss-cpu
sentence-transformers
python-dotenv
